recommend successfully

+++ ServerSide_FinalProject/Assignment1_ServerSide/Models/Series.cs
@@ -15,10 +15,10 @@ namespace Assignment1_ServerSide.Models
         string origin_Country;
         string original_Language;
         string overview;
-        float popularity;
+        double popularity;
         string poster_Path;
 
-        public Series(int id, string first_Air_Date, string name, string origin_Country, string original_Language, string overview, float popularity, string poster_Path)
+        public Series(int id, string first_Air_Date, string name, string origin_Country, string original_Language, string overview, double popularity, string poster_Path)
         {
             Id = id;
             First_Air_Date = first_Air_Date;
@@ -40,7 +40,7 @@ namespace Assignment1_ServerSide.Models
         public string Origin_Country { get => origin_Country; set => origin_Country = value; }
         public string Original_Language { get => original_Language; set => original_Language = value; }
         public string Overview { get => overview; set => overview = value; }
-        public float Popularity { get => popularity; set => popularity = value; }
+        public double Popularity { get => popularity; set => popularity = value; }
         public string Poster_Path { get => poster_Path; set => poster_Path = value; }
